+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/metadata/schema/loader/physical/PhysicalColumnMetaDataLoaderTest.java
@@ -42,7 +42,7 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public final class ColumnMetaDataLoaderTest {
+public final class PhysicalColumnMetaDataLoaderTest {
     
     private static final String TEST_CATALOG = "catalog";
